struggling with sizing tree
